Adding Authorization Checks

+++ controllers/feed.js
@@ -3,6 +3,7 @@ const path = require('path');
 const { validationResult } = require('express-validator/check');
 
 const Post = require('../models/post');
+const User = require('../models/user');
 
 exports.getPosts = (req, res, next) => {
     const currentPage = req.query.page || 1;
@@ -44,22 +45,28 @@ exports.createPost = (req, res, next) => {
     const imageUrl = req.file.path.replace("\\" ,"/");
     const title = req.body.title;
     const content = req.body.content;
+    let creator;
     const post = new Post({
         title: title, 
         content: content,
         imageUrl: imageUrl,
-        creator: {
-            name: 'sotoy'
-        },
-        createdAt: new Date()
+        creator: req.userId,
     });
     post
     .save()
     .then(result => {
-        console.log(result);
+        return User.findById(req.userId)
+    })
+    .then(user => {
+        creator = user;
+        user.posts.push(post);
+        return user.save();
+    })
+    .then(result => {
         res.status(201).json({
             message: 'Successfull',
-            post: result
+            post: post,
+            creator: { _id: creator._id, name: creator.name }
         });
     })
     .catch(err => {
@@ -112,9 +119,14 @@ exports.updatePost = (req, res, next) => {
     }
     Post.findById(postId)
     .then(post => {
-        if(!post) {
+        if (!post) {
             const error = new Error('Could Not find post');
             error.statusCode = 404;
+            throw error;
+        }
+        if (post.creator.toString() !== req.userId) {
+            const error = new Error('Not Authorization');
+            error.statusCode = 403;
             throw error;
         }
         if (imageUrl !== post.imageUrl) {
@@ -148,8 +160,20 @@ exports.deletePost = (req, res, next) => {
             error.statusCode = 404;
             throw error;
         }
+        if (post.creator.toString() !== req.userId) {
+            const error = new Error('Not Authorization');
+            error.statusCode = 403;
+            throw error;
+        }
         clearImage(post.imageUrl);
         return Post.findByIdAndRemove(postId);
+    })
+    .then(() => {
+        return User.findById(req.userId);
+    })
+    .then(user => {
+        user.posts.pull(postId);
+        return user.save()
     })
     .then(result => {
         console.log(result)
